requirements.txt
Flask_Caching==1.7.1
Flask_Script==2.0.6
flask_restplus==0.12.1
Flask==1.0.2
cassandra_driver==3.17.0
